Moved widget and dashboard save logic out of dialog
